Spell check on manual and README correction.
I ran a spell check on the manual and corrected the typos. Also the
introduction for the non-tex related manual output was not correct, so
I fixed that.

The README file needed some editing. I removed the version number from
it because it is very general and not too version related. I also
didn't know any specific and easy way the version number that is set
in `configure.ac` can be included here automatically.